Move new test from Test.Spar to Test.Spar.STM.
