+++ integration/integration.cabal
@@ -170,6 +170,7 @@ library
     Test.Search
     Test.Services
     Test.Spar
+    Test.Spar.STM
     Test.Swagger
     Test.Teams
     Test.TeamSettings
